cleaning/requirements.txt
lxml==4.4.1
asyncpg==0.18.3
psycopg2-binary==2.8.3
pdfminer.six==20181108
chardet==3.0.4
schema==0.7.1
nltk==3.4.5